requirements.txt
# Long Form to Shorts - Version 3.0 Dependencies
# Agentic System Architecture

# Core Dependencies
google-generativeai>=0.8.0  # Gemini API for transcription and analysis
pyyaml>=6.0                  # Configuration management

# Optional but recommended
# rich>=13.0                 # Enhanced CLI output (uncomment to enable)
# tqdm>=4.65                 # Progress bars (uncomment to enable)

# System Requirements (must be installed separately):
# - FFmpeg with libass support: brew install ffmpeg (macOS) or sudo apt install ffmpeg (Linux)
# - Python 3.8 or higher
# - Noto Sans Bengali font for subtitle rendering

# Development Dependencies (optional)
# pytest>=7.0                # Testing framework
# black>=23.0                # Code formatting
# mypy>=1.0                  # Type checking
# pylint>=2.17               # Code linting
